42 step of samurai-way, add redux by MacPook.
