+++ src/App.tsx
@@ -3,7 +3,7 @@ import './App.css';
 import {Header} from './Components/Header/Header';
 import {Footer} from './Components/Footer/Footer';
 import {MainContent} from './Components/MainContent/MainContent';
-import {ActionsTypes, RootStateType} from './Components/redux/state';
+import {ActionsTypes, RootStateType} from './Components/redux/redux-store';
 import {BrowserRouter} from 'react-router-dom';
 
 export type AppType = {
