+++ src/Components/Dialogs/Dialogs.tsx
@@ -2,7 +2,7 @@ import React, {useRef} from 'react';
 import s from './Dialogs.module.css'
 import {DialogItem} from './DialogItem/DialogItem'
 import {Message} from './Message/Message'
-import {ActionsTypes, DialogsPageType} from '../redux/state';
+import {ActionsTypes, DialogsPageType} from '../redux/redux-store';
 import {addMessageAC, updateNewMessageTextAC} from "../redux/dialogs-reducer";
 
 export type DialogPropsType = {
