+++ src/Components/Profile/Profile.tsx
@@ -1,6 +1,6 @@
 import { MyPosts} from './MyPosts/MyPosts';
 import {ProfileInfo} from './ProfileInfo/ProfileInfo';
-import {ActionsTypes, ProfilePageType} from '../redux/state';
+import {ActionsTypes, ProfilePageType} from '../redux/redux-store';
 
 export type ProfilePropsType = {
     profilePage: ProfilePageType
