+++ src/Components/redux/navbar-reducer.ts
@@ -1,4 +1,4 @@
-import {ActionsTypes, RootStateType} from "./state";
+import {ActionsTypes, RootStateType} from "./redux-store";
 
 export const navbarReducer = (state:RootStateType, action: ActionsTypes) => {
     return state
